requirements.txt
beautifulsoup4==4.6.0
boto==2.49.0
boto3==1.10.23
botocore==1.13.23
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-appconf==1.0.3
django-countries==5.3.3
django-crispy-forms==1.8.0
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.1.9
django-ipware==2.1.0
django-mathfilters==0.4.0
django-modelcluster==4.4
django-redis==4.11.0
django-responsive2==0.1.3
django-ses==0.8.12
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.9.4
docutils==0.15.2
draftjs-exporter==2.1.6
future==0.18.2
html5lib==1.0.1
idna==2.8
jmespath==0.9.4
longclaw==1.0.2
oauthlib==3.0.2
Pillow==5.4.1
PyPDF2==1.26.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
razorpay==1.2.0
redis==3.3.11
reportlab==3.5.32
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.3
wagtail==2.5.1
webencodings==0.5.1
Willow==1.1
xhtml2pdf==0.2.3
